% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_data}
\alias{count_data}
\title{Count observations by grouping variables.}
\usage{
count_data(data, grouping_vars, sort_n_desc = FALSE)
}
\arguments{
\item{data}{Input data frame.}

\item{grouping_vars}{A character vector containing grouping variables of a data table.}

\item{sort_n_desc}{A boolean value indicating whether the whole data frame should be sorted in descending order by n. Default is \code{FALSE}.}
}
\value{
A (grouped) data frame.
}
\description{
Count observations by grouping variables.
}
