% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{.ets_pol_id}
\alias{.ets_pol_id}
\title{Generate list of electoral term affilions per politician.}
\usage{
.ets_pol_id(data, id_value)
}
\arguments{
\item{data}{Input \code{data.frame}.}

\item{id_value}{Politician ID values.}
}
\value{
A \code{list}.
}
\description{
Generate list of electoral term affilions per politician.
}
