% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{.group_dummy_fun}
\alias{.group_dummy_fun}
\title{Generate dummy variables for profession variable.}
\usage{
.group_dummy_fun(var, group_name)
}
\arguments{
\item{var}{Name of variable that contains the profession values.}

\item{group_name}{Name of the profession group.}
}
\value{

}
\description{
Generate dummy variables for profession variable.
}
