% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_age}
\alias{get_age}
\title{Get age of politician.}
\usage{
get_age(date_birth, round_val = 2)
}
\arguments{
\item{date_birth}{Birth date of politician.}

\item{round_val}{Numeric value indicating how much the age should be rounded, default is 2.}
}
\value{
Current age of a politician (\code{numeric}).
}
\description{
Get age of politician.
}
