% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_age_hist}
\alias{get_age_hist}
\title{Get age of politician based on a historic date.}
\usage{
get_age_hist(date_hist, date_birth, round_val = 2)
}
\arguments{
\item{date_hist}{Historic date on which age computation is based on.}

\item{date_birth}{Birth date of politician.}

\item{round_val}{Numeric value indicating how much the age should be rounded, default is 2.}
}
\value{
Historic age of a politician based on a certain date, such as the starting date of a legislative period (\code{numeric}).
}
\description{
Get age of politician based on a historic date.
}
