% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_ets}
\alias{get_ets}
\title{Get electoral term affiliations of politicians.}
\usage{
get_ets(data, var, dummy = TRUE, merge = TRUE)
}
\arguments{
\item{data}{Input \code{data.frame}.}

\item{var}{Name of variable (\code{character}) that contains the electoral term values.}

\item{dummy}{A \code{logical} value indicating whether to generate dummy variables per electoral term (\code{TRUE}) or a single \code{list} variable. Default is \code{TRUE}.}

\item{merge}{A \code{logical} value indicating whether to return just the generated columns (\code{FALSE}) or the whole \code{data.frame} (\code{TRUE}). Default is \code{TRUE}.}
}
\value{
A \code{data.frame}.
}
\description{
Get electoral term affiliations of politicians.
}
