% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_faction_color}
\alias{get_faction_color}
\title{Get color code of factions.}
\usage{
get_faction_color(input_id, id_type = "number")
}
\arguments{
\item{input_id}{A \code{character} indicating the ID of the faction.}

\item{id_type}{A \code{character} indicating the type of the \code{input_id}, either \code{"number"} (ID of the faction) or \code{"faction_name"} (abbreviated name of the faction). Default is \code{"number"}.}
}
\value{
A (named) \code{character} yielding a hexadecimal color code.
}
\description{
Get color code of factions.
}
