% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_implausible_data}
\alias{get_implausible_data}
\title{Get data that is implausible based on specific data table based information.}
\usage{
get_implausible_data(data)
}
\arguments{
\item{data}{Input \code{data.frame}.}
}
\value{
A \code{data.frame}.
}
\description{
Get data that is implausible based on specific data table based information.
}
