% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_profession_groups}
\alias{get_profession_groups}
\title{Get professional group affiliation of politicians' jobs.}
\usage{
get_profession_groups(data, var, merge = TRUE)
}
\arguments{
\item{data}{Input data frame.}

\item{var}{Name of variable that contains the profession values.}

\item{merge}{A boolean value indicating whether to return just the profession groups or the whole data frame. Default is \code{TRUE}.}
}
\value{
A data frame.
}
\description{
Get professional group affiliation of politicians' jobs.
}
