% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_state}
\alias{get_state}
\title{Get state affiliation of elected politician (either list or direct mandate).}
\usage{
get_state(politician_id, electoral_term)
}
\arguments{
\item{politician_id}{A \code{character} vector containing the politician ID values.}

\item{electoral_term}{A \code{character} vector containing the electoral term values (number of the legislative period).}
}
\value{
An object of type \code{character}.
}
\description{
Get state affiliation of elected politician (either list or direct mandate).
}
