% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_table_1}
\alias{get_table_1}
\title{Get main table ("Table 1") with descriptive summaries of the database.}
\usage{
get_table_1(table_speeches, table_contributions, output_format = "data.frame")
}
\arguments{
\item{table_speeches}{A \code{data.frame} object, indicating the \code{speeches} table.}

\item{table_contributions}{A \code{data.frame} object, indicating the \code{contributions_simplified} table.}

\item{output_format}{A \code{character}, either \code{"data.frame"} or \code{"latex"}, indicating the output format. Default is \code{"data.frame"}.}
}
\value{
Either a \code{data.frame} (default), or \code{LaTeX} table code.
}
\description{
Get main table ("Table 1") with descriptive summaries of the database.
}
