% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OData.R
\name{OData}
\alias{OData}
\title{Open Discourse Data Object Class}
\description{
Data object class. Provides functionality like retrieving data
or summary functions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{table_name}}{(\code{character}) character string of the table name}

\item{\code{data}}{(\code{dataframe}) data frame of the retrieved table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OData$new()}}
\item \href{#method-get_table_data}{\code{OData$get_table_data()}}
\item \href{#method-get_con}{\code{OData$get_con()}}
\item \href{#method-get_count}{\code{OData$get_count()}}
\item \href{#method-clone}{\code{OData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OData$new(table_name = "factions")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{(\code{character}) character string of the table name
Get Data Table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_table_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_table_data}{}}}
\subsection{Method \code{get_table_data()}}{
Retrieve Data from OD DB and keep it as data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OData$get_table_data(table_name = self$table_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{(\code{character}) character string of the table name
Create db connection from OD docker defaults}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_con"></a>}}
\if{latex}{\out{\hypertarget{method-get_con}{}}}
\subsection{Method \code{get_con()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OData$get_con(
  db = "next",
  host_db = "localhost",
  db_port = "5432",
  db_user = "postgres",
  db_password = "postgres"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{(\code{character} character string of the connection arg)}

\item{\code{host_db}}{(\code{character} character string of the connection arg)}

\item{\code{db_port}}{(\code{character} character string of the connection arg)}

\item{\code{db_user}}{(\code{character} character string of the connection arg)}

\item{\code{db_password}}{(\code{character} character string of the connection arg)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RPostgres conn object
Count observations by grouping variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_count"></a>}}
\if{latex}{\out{\hypertarget{method-get_count}{}}}
\subsection{Method \code{get_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OData$get_count(grouping_vars, sort_n_desc = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grouping_vars}}{A character vector containing grouping variables of a data table.}

\item{\code{sort_n_desc}}{A boolean value indicating whether the whole data frame should be
sorted in descending order by n. Default is \code{FALSE}.}

\item{\code{table_name}}{(\code{character}) character string of the table name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A (grouped) data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
