% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenDiscourse.R
\name{OpenDiscourse}
\alias{OpenDiscourse}
\title{Open Discourse Connector}
\description{
Open Discourse Connector

Open Discourse Connector
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OpenDiscourse$new()}}
\item \href{#method-clone}{\code{OpenDiscourse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenDiscourse$new(docker_config = NULL, dataverse = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{docker_config}}{(\code{list}) docker config as list}

\item{\code{dataverse}}{(\code{logical}) Whether to retrieve data from the harvard
data verse
Connect to docker}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenDiscourse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
