% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_cov}
\alias{plot_cov}
\title{Plotting coverage (\code{NA}s) of data tables.}
\usage{
plot_cov(data, group_var = NULL, exclude_vars = NULL, exclude_na_group = FALSE)
}
\arguments{
\item{data}{Input \code{data.frame}.}

\item{group_var}{\code{character} indicating categorical variable that is used to group the data before plotting.}

\item{exclude_vars}{\code{character} indicating which variable(s) should be excluded from the data before plotting.}

\item{exclude_na_group}{\code{character} indicating whether to include or exclude potential \code{NA}s in the grouping variable. Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plotting coverage (\code{NA}s) of data tables.
}
