% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_rel_freq}
\alias{plot_rel_freq}
\title{Plotting summarized data frames generated by \code{rel_freq_data()}.}
\usage{
plot_rel_freq(data, x_var, fill_var, facet_var = NULL, exclude_na = FALSE)
}
\arguments{
\item{data}{Input \code{data.frame} (function is designed to work with data generated from \code{rel_freq_data()}).}

\item{x_var}{Single \code{character} value indicating which variable is displayed on the x axis.}

\item{fill_var}{Single \code{character} value indicating which variable is used to fill the plot (\code{fill} argument in \code{ggplot()}).}

\item{facet_var}{#' @param facet_var Single \code{character} value indicating which variable is used to facet the plot (via \code{facet_wrap()}) . Default is \code{NULL}.}

\item{exclude_na}{\code{logical} value indicating whether to exclude \code{NA}s in the plot (checks all variables indicated via other parameters). Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plotting summarized data frames generated by \code{rel_freq_data()}.
}
