% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{rel_freq_data}
\alias{rel_freq_data}
\title{Relative frequencies in data when grouping by more than one variable.}
\usage{
rel_freq_data(data, grouping_vars, rel_freq_vars, sort_n_desc = FALSE)
}
\arguments{
\item{data}{Input data frame.}

\item{grouping_vars}{A character vector containing grouping variables of a data table.}

\item{rel_freq_vars}{A character vector containing grouping variables of a data table that have to be part of argument \code{grouping_vars}.}

\item{sort_n_desc}{A boolean value indicating whether the whole data frame should be sorted in descending order by n. Default is \code{FALSE}.}
}
\value{
A (grouped) data frame.
}
\description{
Relative frequencies in data when grouping by more than one variable.
}
